Moved createMultiCallContract to utils and added test
